database/pyproject.toml
[project]
name = "pii-db-tools"
version = "0.1.0"
description = "Database tools and testing utilities for PII encryption system"
authors = [
    {name = "PII Encryption Team"}
]
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[project.scripts]
test-db = "test_connection:main"
run-schema-tests = "run_tests:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
